+++ react/components/RecommendationShelfContainer.tsx
@@ -120,6 +120,7 @@ export const RecommendationShelfContainer: React.FC<Props> = ({
       correlationId={data?.correlationId ?? ''}
       userId={userId}
       displayTitle={displayTitle}
+      campaignId={data?.campaign.id ?? ''}
     />
   ) : (
     <Fragment />