+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Implements ActivityFlow data attibutes
+
 ## [2.13.0] - 2025-09-18
 
 ### Changed