+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- `campaignId` field is now sent on recommendation view and click events.
+
 ## [2.13.0] - 2025-09-18
 
 ### Changed